+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/task/service/RegisterTodoTaskService.java
@@ -3,12 +3,12 @@ package com.LetMeDoWith.LetMeDoWith.application.task.service;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.CreateTodoTaskCommand;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.RegisterTodoTaskResult;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.TodoTaskVO;
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TaskCategoryRepository;
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.common.enums.common.Yn;
 import com.LetMeDoWith.LetMeDoWith.common.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TodoTask;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TaskCategoryRepository;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.task.service.TodoTaskRoutineDateCalculator;
 import java.time.LocalDate;
 import java.util.List;
@@ -19,12 +19,12 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class RegisterTodoTaskService {
-
+    
     private final TodoTaskRepository todoTaskRepository;
     private final TaskCategoryRepository taskCategoryRepository;
     private final TodoTaskRoutineDateCalculator routineDateCalculator;
     private final HolidayFilter holidayFilter;
-
+    
     /**
      * 루틴이 아닌 TodoTask를 생성한다.
      *
@@ -35,19 +35,19 @@ public class RegisterTodoTaskService {
     public RegisterTodoTaskResult registerTodoTask(Long memberId, CreateTodoTaskCommand command) {
         if (command.taskCategoryId() != null) {
             taskCategoryRepository.getActiveTaskCategory(command.taskCategoryId(), memberId)
-                    .orElseThrow(() -> new RestApiException(
-                            FailResponseStatus.DOWITH_TASK_TASK_CATEGORY_NOT_EXIST));
+                                  .orElseThrow(() -> new RestApiException(
+                                      FailResponseStatus.DOWITH_TASK_TASK_CATEGORY_NOT_EXIST));
         }
-
+        
         TodoTask todoTask = TodoTask.of(memberId,
-                command.taskCategoryId(),
-                command.title(),
-                command.startDate(),
-                command.startTime());
-
+                                        command.taskCategoryId(),
+                                        command.title(),
+                                        command.startDate(),
+                                        command.startTime());
+        
         return RegisterTodoTaskResult.of(TodoTaskVO.from(todoTaskRepository.saveTodoTask(todoTask)));
     }
-
+    
     /**
      * TodoTask 루틴을 생성한다.
      *
@@ -57,36 +57,36 @@ public class RegisterTodoTaskService {
      * @return 생성된 루틴의 TodoTask 목록
      */
     public RegisterTodoTaskResult registerTodoTaskRoutine(Long memberId,
-            CreateTodoTaskCommand command) {
+                                                          CreateTodoTaskCommand command) {
         if (command.taskCategoryId() != null) {
             taskCategoryRepository.getTaskCategory(command.taskCategoryId(), Yn.TRUE)
-                    .orElseThrow(() -> new RestApiException(
-                            FailResponseStatus.DOWITH_TASK_TASK_CATEGORY_NOT_EXIST));
+                                  .orElseThrow(() -> new RestApiException(
+                                      FailResponseStatus.DOWITH_TASK_TASK_CATEGORY_NOT_EXIST));
         }
-
+        
         // 루틴 반복 주기에 따른 루틴 수행일자 계산
         Set<LocalDate> routineDates = routineDateCalculator.computeRoutineDates(
-                command.routineCondition().cycle(),
-                command.startDate(),
-                command.endDate(),
-                command.routineCondition().pattern());
-
+            command.routineCondition().cycle(),
+            command.startDate(),
+            command.endDate(),
+            command.routineCondition().pattern());
+        
         // 공휴일 필터 적용
         if (Boolean.TRUE.equals(command.routineCondition().isExcludeHolidays())) {
             routineDates = holidayFilter.filter(routineDates);
         }
-
+        
         List<TodoTask> todoTasks = TodoTask.ofWithRoutine(memberId,
-                command.taskCategoryId(),
-                command.title(),
-                command.startDate(),
-                command.startTime(),
-                routineDates);
-
+                                                          command.taskCategoryId(),
+                                                          command.title(),
+                                                          command.startDate(),
+                                                          command.startTime(),
+                                                          routineDates);
+        
         return RegisterTodoTaskResult.of(todoTaskRepository.saveTodoTasks(todoTasks)
-                .stream()
-                .map(TodoTaskVO::from)
-                .toList(),
-                routineDates);
+                                                           .stream()
+                                                           .map(TodoTaskVO::from)
+                                                           .toList(),
+                                         routineDates);
     }
 }
