+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/task/repository/TaskCategoryRepositoryImpl.java
@@ -1,9 +1,9 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.task.repository;
 
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TaskCategoryRepository;
 import com.LetMeDoWith.LetMeDoWith.common.enums.common.Yn;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TaskCategory;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TaskCategory.TaskCategoryCreationType;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TaskCategoryRepository;
 import com.LetMeDoWith.LetMeDoWith.infrastructure.task.jpaRepository.TaskCategoryJpaRepository;
 import java.util.Arrays;
 import java.util.List;
@@ -14,40 +14,43 @@ import org.springframework.stereotype.Repository;
 @Repository
 @RequiredArgsConstructor
 public class TaskCategoryRepositoryImpl implements TaskCategoryRepository {
-
-  private final TaskCategoryJpaRepository taskCategoryJpaRepository;
-
-  @Override
-  public TaskCategory save(TaskCategory taskCategory) {
-    return taskCategoryJpaRepository.save(taskCategory);
-  }
-
-  @Override
-  public Optional<TaskCategory> getTaskCategory(Long id, Yn isActive) {
-    return taskCategoryJpaRepository.findByIdAndIsActive(id, isActive);
-  }
-
-  @Override
-  public Optional<TaskCategory> getActiveTaskCategory(Long id, Long holderId) {
-    return taskCategoryJpaRepository.findByIdAndCategoryHolderIdInAndIsActive(id,
-        Arrays.asList(holderId, null), Yn.TRUE);
-  }
-
-  @Override
-  public List<TaskCategory> getAllTaskCategories(Yn isActive) {
-    return taskCategoryJpaRepository.findAllByIsActive(isActive);
-  }
-
-  @Override
-  public List<TaskCategory> getCategories(Long holderId, Yn isActive) {
-    return taskCategoryJpaRepository
-        .findAllByCategoryHolderIdAndCreationTypeAndIsActive(holderId,
-            TaskCategoryCreationType.USER_CUSTOM,
-            isActive);
-  }
-
-  @Override
-  public List<TaskCategory> getCategories(TaskCategoryCreationType creationType, Yn isActive) {
-    return taskCategoryJpaRepository.findAllByCreationTypeAndIsActive(creationType, isActive);
-  }
+    
+    private final TaskCategoryJpaRepository taskCategoryJpaRepository;
+    
+    @Override
+    public TaskCategory save(TaskCategory taskCategory) {
+        return taskCategoryJpaRepository.save(taskCategory);
+    }
+    
+    @Override
+    public Optional<TaskCategory> getTaskCategory(Long id, Yn isActive) {
+        return taskCategoryJpaRepository.findByIdAndIsActive(id, isActive);
+    }
+    
+    @Override
+    public Optional<TaskCategory> getActiveTaskCategory(Long id, Long holderId) {
+        return taskCategoryJpaRepository.findByIdAndCategoryHolderIdInAndIsActive(id,
+                                                                                  Arrays.asList(
+                                                                                      holderId,
+                                                                                      null),
+                                                                                  Yn.TRUE);
+    }
+    
+    @Override
+    public List<TaskCategory> getAllTaskCategories(Yn isActive) {
+        return taskCategoryJpaRepository.findAllByIsActive(isActive);
+    }
+    
+    @Override
+    public List<TaskCategory> getCategories(Long holderId, Yn isActive) {
+        return taskCategoryJpaRepository
+            .findAllByCategoryHolderIdAndCreationTypeAndIsActive(holderId,
+                                                                 TaskCategoryCreationType.USER_CUSTOM,
+                                                                 isActive);
+    }
+    
+    @Override
+    public List<TaskCategory> getCategories(TaskCategoryCreationType creationType, Yn isActive) {
+        return taskCategoryJpaRepository.findAllByCreationTypeAndIsActive(creationType, isActive);
+    }
 }
