+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/task/repository/TodoTaskRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.task.repository;
 
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TodoTask;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TodoTaskRoutine;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.infrastructure.task.jpaRepository.TodoTaskJpaRepository;
 import java.time.LocalDate;
 import java.util.List;
