+++ src/test/java/com/LetMeDoWith/LetMeDoWith/application/task/service/RegisterTodoTaskServiceTest.java
@@ -10,14 +10,14 @@ import static org.mockito.Mockito.when;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.CreateTodoTaskCommand;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.CreateTodoTaskCommand.TodoTaskRoutineCondition;
 import com.LetMeDoWith.LetMeDoWith.application.task.dto.RegisterTodoTaskResult;
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TaskCategoryRepository;
-import com.LetMeDoWith.LetMeDoWith.application.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.common.enums.common.Yn;
 import com.LetMeDoWith.LetMeDoWith.common.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.domain.task.enums.TodoTaskRoutineCycle;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TaskCategory;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TodoTask;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TaskCategoryRepository;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.TodoTaskRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.task.service.TodoTaskRoutineDateCalculator;
 import com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy.DailyRoutineDateCalculateStrategy;
 import com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy.TodoTaskRoutineDateCalculateStrategy;
